♻️ refactor: move parser and selector functionality
