+++ src/yutto/downloader/downloader.py
@@ -7,8 +7,8 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 from yutto.bilibili_typing.quality import audio_quality_map, video_quality_map
-from yutto.processor.progressbar import show_progress
-from yutto.processor.selector import select_audio, select_video
+from yutto.downloader.progressbar import show_progress
+from yutto.downloader.selector import select_audio, select_video
 from yutto.utils.asynclib import CoroutineWrapper, first_successful_with_check
 from yutto.utils.console.colorful import colored_string
 from yutto.utils.console.logger import Badge, Logger
