+++ src/yutto/extractor/cheese_batch.py
@@ -7,7 +7,7 @@ from yutto._typing import EpisodeData, EpisodeId, SeasonId
 from yutto.api.cheese import get_cheese_list, get_season_id_by_episode_id
 from yutto.extractor._abc import BatchExtractor
 from yutto.extractor.common import extract_cheese_data
-from yutto.processor.selector import parse_episodes_selection
+from yutto.parser import parse_episodes_selection
 from yutto.utils.asynclib import CoroutineWrapper
 from yutto.utils.console.logger import Badge, Logger
