+++ src/yutto/extractor/ugc_video_batch.py
@@ -8,7 +8,7 @@ from yutto.api.ugc_video import get_ugc_video_list
 from yutto.exceptions import NoAccessPermissionError, NotFoundError
 from yutto.extractor._abc import BatchExtractor
 from yutto.extractor.common import extract_ugc_video_data
-from yutto.processor.selector import parse_episodes_selection
+from yutto.parser import parse_episodes_selection
 from yutto.utils.asynclib import CoroutineWrapper
 from yutto.utils.console.logger import Badge, Logger
