+++ src/yutto/parser.py
@@ -2,71 +2,51 @@ from __future__ import annotations
 
 import re
 import sys
-from typing import TYPE_CHECKING
+import urllib
+import urllib.request
+from pathlib import Path
 
-from yutto.bilibili_typing.codec import (
-    AudioCodec,
-    VideoCodec,
-    gen_acodec_priority,
-    gen_vcodec_priority,
-)
-from yutto.bilibili_typing.quality import (
-    AudioQuality,
-    VideoQuality,
-    gen_audio_quality_priority,
-    gen_video_quality_priority,
-)
 from yutto.exceptions import ErrorCode
 from yutto.utils.console.logger import Logger
 
-if TYPE_CHECKING:
-    from yutto._typing import AudioUrlMeta, VideoUrlMeta
+
+def path_from_cli(path: str) -> Path:
+    """从命令行参数获取路径，支持 ~，以便配置中使用 ~"""
+    return Path(path).expanduser()
 
 
-def select_video(
-    videos: list[VideoUrlMeta],
-    video_quality: VideoQuality = 127,
-    video_codec: VideoCodec = "hevc",
-    video_download_codec_priority: list[VideoCodec] | None = None,
-) -> VideoUrlMeta | None:
-    video_quality_priority = gen_video_quality_priority(video_quality)
-    video_codec_priority = (
-        gen_vcodec_priority(video_codec) if video_download_codec_priority is None else video_download_codec_priority
-    )
-
-    video_combined_priority = [
-        (vqn, vcodec)
-        for vqn in video_quality_priority
-        # TODO: Dolby Selector
-        for vcodec in video_codec_priority
-    ]  # fmt: skip
-
-    for vqn, vcodec in video_combined_priority:
-        for video in videos:
-            if video["quality"] == vqn and video["codec"] == vcodec:
-                return video
-    return None
+def is_comment(line: str) -> bool:
+    """判断文件某行是否为注释"""
+    if line.startswith("#"):
+        return True
+    return False
 
 
-def select_audio(
-    audios: list[AudioUrlMeta],
-    audio_quality: AudioQuality = 30280,
-    audio_codec: AudioCodec = "mp4a",
-) -> AudioUrlMeta | None:
-    audio_quality_priority = gen_audio_quality_priority(audio_quality)
-    audio_codec_priority = gen_acodec_priority(audio_codec)
+def alias_parser(file_path: str) -> dict[str, str]:
+    result: dict[str, str] = {}
+    re_alias_splitter = re.compile(r"[\s=]")
+    with path_from_cli(file_path).open("r") as f_alias:
+        for line in f_alias:
+            line = line.strip()
+            if not line or is_comment(line):
+                continue
+            alias, url = re_alias_splitter.split(line, maxsplit=1)
+            result[alias] = url
+    return result
 
-    audio_combined_priority = [
-        (aqn, acodec)
-        for aqn in audio_quality_priority
-        for acodec in audio_codec_priority
-    ]  # fmt: skip
 
-    for aqn, acodec in audio_combined_priority:
-        for audio in audios:
-            if audio["quality"] == aqn and audio["codec"] == acodec:
-                return audio
-    return None
+def file_scheme_parser(url: str) -> list[str]:
+    file_url: str = urllib.parse.urlparse(url).path  # type: ignore
+    file_path = path_from_cli(urllib.request.url2pathname(file_url))  # type: ignore
+    Logger.info(f"解析下载列表 {file_path} 中...")
+    result: list[str] = []
+    with file_path.open("r", encoding="utf-8") as f:
+        for line in f:
+            line = line.strip()
+            if not line or is_comment(line):
+                continue
+            result.append(line)
+    return result
 
 
 def validate_episodes_selection(episodes_str: str) -> bool:
