+++ src/yutto/validator.py
@@ -11,7 +11,7 @@ import biliass
 from yutto.api.user_info import get_user_info
 from yutto.bilibili_typing.codec import VideoCodec, audio_codec_priority_default, video_codec_priority_default
 from yutto.exceptions import ErrorCode
-from yutto.processor.selector import validate_episodes_selection
+from yutto.parser import validate_episodes_selection
 from yutto.utils.asynclib import initial_async_policy
 from yutto.utils.console.colorful import set_no_color
 from yutto.utils.console.logger import Badge, Logger, set_logger_debug
